refactor(core)!: split presentation into fullscreen and pip features
Split PresentationState into FullscreenState and PictureInPictureState,
each with its own feature, selector, and tests.

- Rename fullscreenActive -> fullscreen, pipActive -> pip
- presentationFeature -> fullscreenFeature + pipFeature
- selectPresentation -> selectFullscreen + selectPiP
- features.presentation -> features.fullscreen + features.pip
- Video preset now includes both features separately
