+++ packages/core/src/dom/store/features/feature.parts.ts
@@ -1,6 +1,7 @@
 import { bufferFeature } from './buffer';
+import { fullscreenFeature } from './fullscreen';
+import { pipFeature } from './pip';
 import { playbackFeature } from './playback';
-import { presentationFeature } from './presentation';
 import { sourceFeature } from './source';
 import { timeFeature } from './time';
 import { volumeFeature } from './volume';
@@ -8,8 +9,9 @@ import { volumeFeature } from './volume';
 // Short aliases
 export {
   bufferFeature as buffer,
+  fullscreenFeature as fullscreen,
+  pipFeature as pip,
   playbackFeature as playback,
-  presentationFeature as presentation,
   sourceFeature as source,
   timeFeature as time,
   volumeFeature as volume,
@@ -21,7 +23,8 @@ export const video = [
   timeFeature,
   sourceFeature,
   bufferFeature,
-  presentationFeature,
+  fullscreenFeature,
+  pipFeature,
 ] as const;
 
 export const audio = [playbackFeature, volumeFeature, timeFeature, sourceFeature, bufferFeature] as const;
