+++ packages/core/src/dom/store/features/index.ts
@@ -1,7 +1,8 @@
 export * from './buffer';
 export * as features from './feature.parts';
+export * from './fullscreen';
+export * from './pip';
 export * from './playback';
-export * from './presentation';
 export * from './source';
 export * from './time';
 export * from './volume';
