+++ packages/core/src/dom/store/features/tests/fullscreen.test.ts
@@ -2,36 +2,22 @@ import { createStore } from '@videojs/store';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import type { PlayerTarget } from '../../../media/types';
-import { presentationFeature } from '../presentation';
+import { fullscreenFeature } from '../fullscreen';
 
-describe('presentationFeature', () => {
+describe('fullscreenFeature', () => {
   let originalFullscreenEnabled: boolean | undefined;
-  let originalPictureInPictureEnabled: boolean | undefined;
 
   beforeEach(() => {
-    // Save original values
     originalFullscreenEnabled = document.fullscreenEnabled;
-    originalPictureInPictureEnabled = document.pictureInPictureEnabled;
   });
 
   afterEach(() => {
-    // Restore original values
     Object.defineProperty(document, 'fullscreenEnabled', {
       value: originalFullscreenEnabled,
       writable: true,
       configurable: true,
     });
-    Object.defineProperty(document, 'pictureInPictureEnabled', {
-      value: originalPictureInPictureEnabled,
-      writable: true,
-      configurable: true,
-    });
     Object.defineProperty(document, 'fullscreenElement', {
-      value: null,
-      writable: true,
-      configurable: true,
-    });
-    Object.defineProperty(document, 'pictureInPictureElement', {
       value: null,
       writable: true,
       configurable: true,
@@ -43,11 +29,10 @@ describe('presentationFeature', () => {
       const video = createMockVideo();
       const container = document.createElement('div');
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container });
 
-      expect(store.state.fullscreenActive).toBe(false);
-      expect(store.state.pipActive).toBe(false);
+      expect(store.state.fullscreen).toBe(false);
     });
 
     it('detects fullscreen availability when supported', () => {
@@ -58,7 +43,7 @@ describe('presentationFeature', () => {
       });
 
       const video = createMockVideo();
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container: null });
 
       expect(store.state.fullscreenAvailability).toBe('available');
@@ -72,27 +57,13 @@ describe('presentationFeature', () => {
       });
 
       const video = createMockVideo();
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container: null });
 
       expect(store.state.fullscreenAvailability).toBe('unsupported');
     });
 
-    it('detects PiP availability when supported', () => {
-      Object.defineProperty(document, 'pictureInPictureEnabled', {
-        value: true,
-        writable: true,
-        configurable: true,
-      });
-
-      const video = createMockVideo();
-      const store = createStore<PlayerTarget>()(presentationFeature);
-      store.attach({ media: video, container: null });
-
-      expect(store.state.pipAvailability).toBe('available');
-    });
-
-    it('updates fullscreenActive on fullscreenchange event', () => {
+    it('updates fullscreen on fullscreenchange event', () => {
       Object.defineProperty(document, 'fullscreenEnabled', {
         value: true,
         writable: true,
@@ -102,10 +73,10 @@ describe('presentationFeature', () => {
       const video = createMockVideo();
       const container = document.createElement('div');
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container });
 
-      expect(store.state.fullscreenActive).toBe(false);
+      expect(store.state.fullscreen).toBe(false);
 
       // Simulate entering fullscreen
       Object.defineProperty(document, 'fullscreenElement', {
@@ -115,7 +86,7 @@ describe('presentationFeature', () => {
       });
       document.dispatchEvent(new Event('fullscreenchange'));
 
-      expect(store.state.fullscreenActive).toBe(true);
+      expect(store.state.fullscreen).toBe(true);
 
       // Simulate exiting fullscreen
       Object.defineProperty(document, 'fullscreenElement', {
@@ -125,42 +96,7 @@ describe('presentationFeature', () => {
       });
       document.dispatchEvent(new Event('fullscreenchange'));
 
-      expect(store.state.fullscreenActive).toBe(false);
-    });
-
-    it('updates pipActive on PiP events', () => {
-      Object.defineProperty(document, 'pictureInPictureEnabled', {
-        value: true,
-        writable: true,
-        configurable: true,
-      });
-
-      const video = createMockVideo();
-
-      const store = createStore<PlayerTarget>()(presentationFeature);
-      store.attach({ media: video, container: null });
-
-      expect(store.state.pipActive).toBe(false);
-
-      // Simulate entering PiP
-      Object.defineProperty(document, 'pictureInPictureElement', {
-        value: video,
-        writable: true,
-        configurable: true,
-      });
-      video.dispatchEvent(new Event('enterpictureinpicture'));
-
-      expect(store.state.pipActive).toBe(true);
-
-      // Simulate exiting PiP
-      Object.defineProperty(document, 'pictureInPictureElement', {
-        value: null,
-        writable: true,
-        configurable: true,
-      });
-      video.dispatchEvent(new Event('leavepictureinpicture'));
-
-      expect(store.state.pipActive).toBe(false);
+      expect(store.state.fullscreen).toBe(false);
     });
 
     it('stops listening when store is destroyed', () => {
@@ -173,7 +109,7 @@ describe('presentationFeature', () => {
       const video = createMockVideo();
       const container = document.createElement('div');
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container });
 
       store.destroy();
@@ -187,7 +123,7 @@ describe('presentationFeature', () => {
       document.dispatchEvent(new Event('fullscreenchange'));
 
       // State should not update after destroy
-      expect(store.state.fullscreenActive).toBe(false);
+      expect(store.state.fullscreen).toBe(false);
     });
   });
 
@@ -197,7 +133,7 @@ describe('presentationFeature', () => {
       const container = document.createElement('div');
       container.requestFullscreen = vi.fn().mockResolvedValue(undefined);
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container });
 
       await store.requestFullscreen();
@@ -209,7 +145,7 @@ describe('presentationFeature', () => {
       const video = createMockVideo();
       video.requestFullscreen = vi.fn().mockResolvedValue(undefined);
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container: null });
 
       await store.requestFullscreen();
@@ -223,7 +159,7 @@ describe('presentationFeature', () => {
 
       const video = createMockVideo();
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container: null });
 
       await store.exitFullscreen();
@@ -231,41 +167,6 @@ describe('presentationFeature', () => {
       expect(document.exitFullscreen).toHaveBeenCalled();
 
       document.exitFullscreen = originalExit;
-    });
-
-    it('requestPiP() calls requestPictureInPicture on video', async () => {
-      const video = createMockVideo();
-      video.requestPictureInPicture = vi.fn().mockResolvedValue({});
-
-      const store = createStore<PlayerTarget>()(presentationFeature);
-      store.attach({ media: video, container: null });
-
-      await store.requestPiP();
-
-      expect(video.requestPictureInPicture).toHaveBeenCalled();
-    });
-
-    it('exitPiP() calls document.exitPictureInPicture', async () => {
-      const originalExit = document.exitPictureInPicture;
-      document.exitPictureInPicture = vi.fn().mockResolvedValue(undefined);
-
-      const video = createMockVideo();
-
-      // Set the video as the current PiP element
-      Object.defineProperty(document, 'pictureInPictureElement', {
-        value: video,
-        writable: true,
-        configurable: true,
-      });
-
-      const store = createStore<PlayerTarget>()(presentationFeature);
-      store.attach({ media: video, container: null });
-
-      await store.exitPiP();
-
-      expect(document.exitPictureInPicture).toHaveBeenCalled();
-
-      document.exitPictureInPicture = originalExit;
     });
   });
 
@@ -285,7 +186,7 @@ describe('presentationFeature', () => {
         configurable: true,
       });
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container });
 
       await store.requestFullscreen();
@@ -296,32 +197,6 @@ describe('presentationFeature', () => {
       document.exitPictureInPicture = originalExit;
     });
 
-    it('requestPiP() exits fullscreen first if active', async () => {
-      const originalExit = document.exitFullscreen;
-      document.exitFullscreen = vi.fn().mockResolvedValue(undefined);
-
-      const video = createMockVideo();
-      video.requestPictureInPicture = vi.fn().mockResolvedValue({});
-      const container = document.createElement('div');
-
-      // Set fullscreen as active
-      Object.defineProperty(document, 'fullscreenElement', {
-        value: container,
-        writable: true,
-        configurable: true,
-      });
-
-      const store = createStore<PlayerTarget>()(presentationFeature);
-      store.attach({ media: video, container });
-
-      await store.requestPiP();
-
-      expect(document.exitFullscreen).toHaveBeenCalled();
-      expect(video.requestPictureInPicture).toHaveBeenCalled();
-
-      document.exitFullscreen = originalExit;
-    });
-
     it('requestFullscreen() does not exit PiP if not active', async () => {
       const originalExit = document.exitPictureInPicture;
       document.exitPictureInPicture = vi.fn().mockResolvedValue(undefined);
@@ -330,7 +205,7 @@ describe('presentationFeature', () => {
       const container = document.createElement('div');
       container.requestFullscreen = vi.fn().mockResolvedValue(undefined);
 
-      const store = createStore<PlayerTarget>()(presentationFeature);
+      const store = createStore<PlayerTarget>()(fullscreenFeature);
       store.attach({ media: video, container });
 
       await store.requestFullscreen();
@@ -339,24 +214,6 @@ describe('presentationFeature', () => {
       expect(container.requestFullscreen).toHaveBeenCalled();
 
       document.exitPictureInPicture = originalExit;
-    });
-
-    it('requestPiP() does not exit fullscreen if not active', async () => {
-      const originalExit = document.exitFullscreen;
-      document.exitFullscreen = vi.fn().mockResolvedValue(undefined);
-
-      const video = createMockVideo();
-      video.requestPictureInPicture = vi.fn().mockResolvedValue({});
-
-      const store = createStore<PlayerTarget>()(presentationFeature);
-      store.attach({ media: video, container: null });
-
-      await store.requestPiP();
-
-      expect(document.exitFullscreen).not.toHaveBeenCalled();
-      expect(video.requestPictureInPicture).toHaveBeenCalled();
-
-      document.exitFullscreen = originalExit;
     });
   });
 });
