+++ packages/core/src/dom/store/selectors.ts
@@ -1,15 +1,17 @@
 import { createSelector } from '@videojs/store';
 
 import { bufferFeature } from './features/buffer';
+import { fullscreenFeature } from './features/fullscreen';
+import { pipFeature } from './features/pip';
 import { playbackFeature } from './features/playback';
-import { presentationFeature } from './features/presentation';
 import { sourceFeature } from './features/source';
 import { timeFeature } from './features/time';
 import { volumeFeature } from './features/volume';
 
 export const selectBuffer = createSelector(bufferFeature);
+export const selectFullscreen = createSelector(fullscreenFeature);
+export const selectPiP = createSelector(pipFeature);
 export const selectPlayback = createSelector(playbackFeature);
-export const selectPresentation = createSelector(presentationFeature);
 export const selectSource = createSelector(sourceFeature);
 export const selectTime = createSelector(timeFeature);
 export const selectVolume = createSelector(volumeFeature);
